fix: use html-minifier-terser instead of regex. feat: add option to skip pre-processing (#51)
* feat: add option to skip pre-processing

* fix: remove readme issue

* feat: initial implementation of html-minifier-terser

* chore: terser improvements
