+++ README.md
@@ -81,6 +81,8 @@ full fledged examples can be found under `example/`
     - `defaultOrderedListStyleType` <?[String]> default ordered list style type. Defaults to `decimal`.
   - `decodeUnicode` <?[Boolean]> flag to enable unicode decoding of header, body and footer. Defaults to `false`.
   - `lang` <?[String]> language localization code for spell checker to work properly. Defaults to `en-US`.
+  - `preProcessing` <?[Object]>
+    - `skipHTMLMinify` <?[Boolean]> flag to skip minification of HTML. Defaults to `false`.
 - `footerHTMLString` <[String]> clean html string equivalent of footer. Defaults to `<p></p>` if footer flag is `true`.
 
 ### Returns
